Moved nupic.regions.Spec -> nupic.regions.spec
