+++ src/nupic/regions/spec.py
